organizing utilities into directories
