+++ dvh/__main__.py
@@ -8,12 +8,12 @@ This is the main python file for command line implementation.
 
 
 from __future__ import print_function
-import update_sys_path  # noqa
 from os.path import dirname, isfile, isdir, join
-from tools.dicom_to_sql import dicom_to_sql
-from tools.sql_connector import DVH_SQL
-from tools.utilities import is_import_settings_defined, is_sql_connection_defined, write_import_settings,\
-    write_sql_connection_settings, validate_sql_connection, initialize_directories_settings
+from tools.io.database.dicom.importer import dicom_to_sql
+from tools.io.database.sql_connector import DVH_SQL
+from tools.io.preferences.import_settings import is_import_settings_defined, write_import_settings,\
+    initialize_directories_settings
+from tools.io.preferences.sql import is_sql_connection_defined, write_sql_connection_settings, validate_sql_connection
 from tools.get_settings import get_settings
 from getpass import getpass
 import argparse
