+++ dvh/dvh_bokeh_models/main/rad_bio.py
@@ -9,7 +9,7 @@ from bokeh.models.widgets import Button, TextInput, CheckboxButtonGroup, Div
 from bokeh.models import Spacer
 from bokeh.layouts import column, row
 from tools.utilities import get_include_map
-from tools.analysis_tools import calc_eud
+from tools.io.database.analysis_tools import calc_eud
 import options
 from options import GROUP_LABELS
