+++ dvh/tools/get_settings.py
@@ -1,13 +1,5 @@
-import sys
-from os.path import isfile, dirname, join, realpath
-sys.path.append(dirname(dirname(realpath(__file__))))
-from paths import PREF_DIR
-
-
-SETTINGS_PATHS = {'docker': {'import': "/import_and_connection/import_settings.txt",
-                             'sql': "/import_and_connection/sql_connection.cnf"},
-                  'default': {'import': join(PREF_DIR, 'import_settings.txt'),
-                              'sql': join(PREF_DIR, 'sql_connection.cnf')}}
+from os.path import isfile
+from options import SETTINGS_PATHS
 
 
 def get_settings(settings_type):
