+++ dvh/tools/io/database/sql_to_python.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import print_function
-from sql_connector import DVH_SQL
+from tools.io.database.sql_connector import DVH_SQL
 
 
 class QuerySQL:
