+++ dvh/tools/roi/geometry.py
@@ -1,46 +1,11 @@
-from shapely.geometry import Polygon, Point
-from shapely import speedups
 from scipy.spatial.distance import cdist
 import numpy as np
+from math import ceil
+from tools.roi.formatter import points_to_shapely_polygon, dicompyler_roi_to_sets_of_points,\
+    get_shapely_from_sets_of_points
 
 
-MIN_SLICE_THICKNESS = 2  # Update method to pull from DICOM
-
-
-# Enable shapely calculations using C, as opposed to the C++ default
-if speedups.available:
-    speedups.enable()
-
-
-def get_planes_from_string(roi_coord_string):
-    """
-    :param roi_coord_string: roi string represntation of an roi as formatted in the SQL database
-    :return: a "sets of points" formatted list
-    :rtype: list
-    """
-    planes = {}
-    contours = roi_coord_string.split(':')
-
-    for contour in contours:
-        contour = contour.split(',')
-        z = contour.pop(0)
-        z = round(float(z), 2)
-        z_str = str(z)
-
-        if z_str not in list(planes):
-            planes[z_str] = []
-
-        i, points = 0, []
-        while i < len(contour):
-            point = [float(contour[i]), float(contour[i+1]), z]
-            points.append(point)
-            i += 2
-        planes[z_str].append(points)
-
-    return planes
-
-
-def get_union(rois):
+def union(rois):
     """
     :param rois: a list of "sets of points"
     :return: a "sets of points" representing the union of the rois, each item in "sets of points" is a plane
@@ -95,75 +60,7 @@ def get_union(rois):
     return new_roi
 
 
-def points_to_shapely_polygon(sets_of_points):
-    """
-    :param sets_of_points: sets of points is a dictionary of lists using str(z) as keys
-    :return: a composite polygon as a shapely object (eith polygon or multipolygon)
-    """
-    # sets of points are lists using str(z) as keys
-    # each item is an ordered list of points representing a polygon, each point is a 3-item list [x, y, z]
-    # polygon n is inside polygon n-1, then the current accumulated polygon is
-    #    polygon n subtracted from the accumulated polygon up to and including polygon n-1
-    #    Same method DICOM uses to handle rings and islands
-
-    composite_polygon = []
-    for set_of_points in sets_of_points:
-        if len(set_of_points) > 3:
-            points = [(point[0], point[1]) for point in set_of_points]
-            points.append(points[0])  # Explicitly connect the final point to the first
-
-            # if there are multiple sets of points in a slice, each set is a polygon,
-            # interior polygons are subtractions, exterior are addition
-            # Only need to check one point for interior vs exterior
-            current_polygon = Polygon(points).buffer(0)  # clean stray points
-            if composite_polygon:
-                if Point((points[0][0], points[0][1])).disjoint(composite_polygon):
-                    composite_polygon = composite_polygon.union(current_polygon)
-                else:
-                    composite_polygon = composite_polygon.symmetric_difference(current_polygon)
-            else:
-                composite_polygon = current_polygon
-
-    return composite_polygon
-
-
-def get_roi_coordinates_from_string(roi_coord_string):
-    """
-    :param roi_coord_string: the string reprentation of an roi in the SQL database
-    :return: a list of numpy arrays, each array is the x, y, z coordinates of the given point
-    :rtype: list
-    """
-    roi_coordinates = []
-    contours = roi_coord_string.split(':')
-
-    for contour in contours:
-        contour = contour.split(',')
-        z = contour.pop(0)
-        z = float(z)
-        i = 0
-        while i < len(contour):
-            roi_coordinates.append(np.array((float(contour[i]), float(contour[i + 1]), z)))
-            i += 2
-
-    return roi_coordinates
-
-
-def get_roi_coordinates_from_planes(planes):
-    """
-    :param planes: a "sets of points" formatted list
-    :return: a list of numpy arrays, each array is the x, y, z coordinates of the given point
-    :rtype: list
-    """
-    roi_coordinates = []
-
-    for z in list(planes):
-        for polygon in planes[z]:
-            for point in polygon:
-                roi_coordinates.append(np.array((point[0], point[1], point[2])))
-    return roi_coordinates
-
-
-def get_min_distances_to_target(oar_coordinates, target_coordinates):
+def min_distances_to_target(oar_coordinates, target_coordinates):
     """
     :param oar_coordinates: list of numpy arrays of 3D points defining the surface of the OAR
     :param target_coordinates: list of numpy arrays of 3D points defining the surface of the PTV
@@ -178,7 +75,7 @@ def get_min_distances_to_target(oar_coordinates, target_coordinates):
     return min_distances
 
 
-def calc_cross_section(roi):
+def cross_section(roi):
     """
     :param roi: a "sets of points" formatted list
     :return: max and median cross-sectional area of all slices in cm^2
@@ -205,24 +102,7 @@ def calc_cross_section(roi):
     return area
 
 
-def dicompyler_roi_coord_to_db_string(coord):
-    """
-    :param coord: dicompyler structure coordinates from GetStructureCoordinates()
-    :return: string representation of roi, <z1>: <x1 y1 x2 y2... xn yn>, <zn>: <x1 y1 x2 y2... xn yn>
-    :rtype: str
-    """
-    contours = []
-    for z in coord:
-        for plane in coord[z]:
-            points = [z]
-            for point in plane['data']:
-                points.append(str(round(point[0], 3)))
-                points.append(str(round(point[1], 3)))
-            contours.append(','.join(points))
-    return ':'.join(contours)
-
-
-def surface_area_of_roi(coord, coord_type='dicompyler'):
+def surface_area(coord, coord_type='dicompyler'):
     """
     :param coord: dicompyler structure coordinates from GetStructureCoordinates() or sets_of_points
     :param coord_type: either 'dicompyler' or 'sets_of_points'
@@ -258,56 +138,7 @@ def surface_area_of_roi(coord, coord_type='dicompyler'):
     return round(area/100, 3)
 
 
-def get_shapely_from_sets_of_points(sets_of_points):
-    """
-    :param sets_of_points: a dictionary of slices with key being a str representation of z value, value is a list
-    of points defining a polygon in the slice.  point[0] is x and point[1] is y
-    :return: roi_slice which is a dictionary of lists of z, thickness, and a Shapely Polygon class object
-    :rtype: list
-    """
-
-    roi_slice = {'z': [], 'thickness': [], 'polygon': []}
-
-    sets_of_points_keys = list(sets_of_points)
-    sets_of_points_keys.sort()
-
-    all_z_values = [round(float(z), 2) for z in sets_of_points_keys]
-    thicknesses = np.abs(np.diff(all_z_values))
-    if len(thicknesses):
-        thicknesses = np.append(thicknesses, np.min(thicknesses))
-    else:
-        thicknesses = np.array([MIN_SLICE_THICKNESS])
-
-    for z in sets_of_points:
-        thickness = thicknesses[all_z_values.index(round(float(z), 2))]
-        shapely_roi = points_to_shapely_polygon(sets_of_points[z])
-        if shapely_roi:
-            roi_slice['z'].append(round(float(z), 2))
-            roi_slice['thickness'].append(thickness)
-            roi_slice['polygon'].append(shapely_roi)
-
-    return roi_slice
-
-
-def dicompyler_roi_to_sets_of_points(coord):
-    """
-    :param coord: dicompyler structure coordinates from GetStructureCoordinates()
-    :return: a dictionary of lists of points that define contours in each slice z
-    :rtype: dict
-    """
-    all_points = {}
-    for z in coord:
-        all_points[z] = []
-        for plane in coord[z]:
-            plane_points = [[float(point[0]), float(point[1])] for point in plane['data']]
-            for point in plane['data']:
-                plane_points.append([float(point[0]), float(point[1])])
-            if len(plane_points) > 2:
-                all_points[z].append(plane_points)
-    return all_points
-
-
-def calc_roi_overlap(oar, tv):
+def overlap_volume(oar, tv):
     """
     :param oar: dict representing organ-at-risk, follows format of "sets of points" in dicompyler_roi_to_sets_of_points
     :param tv: dict representing treatment volume
@@ -336,7 +167,7 @@ def calc_roi_overlap(oar, tv):
     return round(intersection_volume / 1000., 2)
 
 
-def calc_volume(roi):
+def volume(roi):
     """
     :param roi: a "sets of points" formatted list
     :return: volume in cm^3 of roi
@@ -349,7 +180,7 @@ def calc_volume(roi):
     #    polygon n subtracted from the accumulated polygon up to and including polygon n-1
     #    Same method DICOM uses to handle rings and islands
 
-    volume = 0.
+    vol = 0.
     all_z_values = [round(float(z), 2) for z in list(roi)]
     all_z_values = np.sort(all_z_values)
     thicknesses = np.abs(np.diff(all_z_values))
@@ -363,12 +194,12 @@ def calc_volume(roi):
         thickness = thicknesses[all_z_values.index(round(float(z), 2))]
         shapely_roi = points_to_shapely_polygon(roi[z])
         if shapely_roi:
-            volume += shapely_roi.area * thickness
+            vol += shapely_roi.area * thickness
 
-    return round(volume / 1000., 2)
+    return round(vol / 1000., 2)
 
 
-def calc_centroid(roi):
+def centroid(roi):
     """
     :param roi: a "sets of points" formatted list
     :return: centroid dicom coordinate in mm
@@ -396,8 +227,47 @@ def calc_centroid(roi):
     w = np.array(centroids['area'])
     w_sum = np.sum(w)
 
-    centroid = [float(np.sum(x * w) / w_sum),
-                float(np.sum(y * w) / w_sum),
-                float(np.sum(z * w) / w_sum)]
+    volumetric_centroid = [float(np.sum(x * w) / w_sum),
+                           float(np.sum(y * w) / w_sum),
+                           float(np.sum(z * w) / w_sum)]
 
-    return centroid
+    return volumetric_centroid
+
+
+def spread(roi):
+    """
+    :param roi: a "sets of points" formatted list
+    :return: x, y, z dimensions of a rectangular prism encompassing roi
+    :rtype: list
+    """
+    all_points = {'x': [], 'y': [], 'z': []}
+
+    for z in list(roi):
+        for polygon in roi[z]:
+            for point in polygon:
+                all_points['x'].append(point[0])
+                all_points['y'].append(point[1])
+                all_points['z'].append(point[2])
+
+    all_points = {'x': np.array(all_points['x']),
+                  'y': np.array(all_points['y']),
+                  'z': np.array(all_points['z'])}
+
+    if len(all_points['x'] > 1):
+        data = [abs(float(np.max(all_points[dim]) - np.min(all_points[dim]))) for dim in ['x', 'y', 'z']]
+    else:
+        data = [0, 0, 0]
+
+    return data
+
+
+def dth(min_distances):
+    """
+    :param min_distances:
+    :return: histogram of distances in 0.1mm bin widths
+    """
+
+    bin_count = int(ceil(np.max(min_distances) * 10.))
+    data, bins = np.histogram(min_distances, bins=bin_count)
+
+    return data
